services/log-service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
